Split hold note blueprints into separate file
